Infinite chunk loading (+ varint fixes) (#11)
* Fixed a bunch of bugs, made an attempt on infinite chunks, but the client doesn't seem to render them.

* Infinite chunk loading works, fixed negative varints.
TODO: Fix a crash that occurs very often when reading the packet batch length. It reads the unsigned varInt incorrectly under some conditions, causing the server to crash because the length read is way longer than the actual length.

* Fixed crash if uninitialized player would join, fixed very big size varints, fixed split packets with lots of fragments missing buffers.

+++ src/gomine/interfaces/IChunk.go
@@ -27,7 +27,7 @@ type IChunk interface {
 	GetSubChunks() map[int]ISubChunk
 	GetHighestBlockId(int, int) byte
 	GetHighestBlockData(int, int) byte
-	GetHighestBlock(int, int) int
+	GetHighestBlock(int, int) int16
 	ToBinary() []byte
 	RecalculateHeightMap()
 }
